+++ app/controllers/preview/entries_controller.rb
@@ -1,6 +1,11 @@
 class Preview::EntriesController < ApplicationController
   def show
-    @journey = Journey.create(category: "catering", liquid_template: "<p>N/A</p>")
+    @journey = Journey.create(
+      category: "catering",
+      user: current_user,
+      liquid_template: "<p>N/A</p>"
+    )
+
     contentful_entry = GetEntry.new(entry_id: entry_id).call
     @step = CreateJourneyStep.new(
       journey: @journey, contentful_entry: contentful_entry